+++ CHANGELOG.rst
@@ -10,6 +10,9 @@ Unreleased
 New Features
 ~~~~~~~~~~~~
 
+* util-core: `c.t.conversions` now includes conversion methods `toOption`, and `getOrElse`
+  under `StringOps`. ``PHAB_ID=D578549``
+
 * util-core: `c.t.util.Duration` now includes `fromJava` and `asJava` conversions to
   `java.time.Duration` types. ``PHAB_ID=D571885``
 
@@ -31,7 +34,6 @@ New Features
 
 * util-stats: `c.t.finagle.stats.DenylistStatsReceiver` now includes methods for creating
   `DenyListStatsReceiver` from partial functions. ``PHAB_ID=D576833``
-
 
 20.10.0
 -------